adds modal to case dist contect, adds COPY, moves save button to modal, updates leverbuttons (#19919)
