+++ client/COPY.json
@@ -1407,5 +1407,7 @@
   "CASE_DISTRIBUTION_STATIC_LEVERS_TITLE": "Inactive Data Elements",
   "CASE_DISTRIBUTION_STATIC_LEVERS_DESCRIPTION": "The below values do not effect the current active Case Distribution algorithm.",
   "CASE_DISTRIBUTION_HISTORY_TITLE": "Case Distribution Algorithm Change History",
-  "CASE_DISTRIBUTION_HISTORY_DESCRIPTION": "Change history for Case Distribution Admin shows changes made in the last 365 days."
+  "CASE_DISTRIBUTION_HISTORY_DESCRIPTION": "Change history for Case Distribution Admin shows changes made in the last 365 days.",
+  "CASE_DISTRIBUTION_MODAL_TITLE": "Confirm Case Distribution Algorithm Changes",
+  "CASE_DISTRIBUTION_MODAL_DESCRIPTION": "Please confirm the following changes to the Case Distribution algorithm below. Once saved, these changes will apply to the next scheduled case distribution event."
 }
